added execution and stock buyying nad selling execution controller provides buy and sell option with buy having symbol and price and sell having symbol and quantity also added their subsequent dto and also a shar edto to get the result from the tradeing api and process it. stock repo stores stocks and now also increments and decrements stock price. and price fetch and execution service do the logic for the buy and sell. finllay solved a bug in dashboard where the total amount was misscaculated

+++ src/main/java/majorproject/maf/repository/StockRepository.java
@@ -19,7 +19,7 @@ public interface StockRepository extends JpaRepository<Stock, String> {
         SET h.shares = h.shares + :delta
         WHERE h.user.id = :userId AND h.symbol = :symbol
     """)
-    int incrementShares(int userId, String symbol, int delta);
+    int incrementShares(int userId, String symbol, double delta);
 
     @Modifying
     @Query("""
@@ -27,5 +27,5 @@ public interface StockRepository extends JpaRepository<Stock, String> {
         SET h.shares = h.shares - :delta
         WHERE h.user.id = :userId AND h.symbol = :symbol
     """)
-    int decrementShares(int userId, String symbol, int delta);
+    int decrementShares(int userId, String symbol, double delta);
 }
